Move `intermediate.rendering` to `doc`
We merge the module `intermediate.rendering` into `intermediate.doc` as
it is directly related to it, and no other functionalities emerged in
`intermediate.rendering`.

Originally, we thought that `intermediate.rendering` would span more
functionalities, which did not materialize.
