+++ aas_core_codegen/csharp/description.py
@@ -36,7 +36,6 @@ from aas_core_codegen.csharp import (
 )
 from aas_core_codegen.intermediate import (
     doc as intermediate_doc,
-    rendering as intermediate_rendering,
     _translate as intermediate_translate,
 )
 
@@ -153,7 +152,7 @@ class _NodeVisitor:
         self.visit(node.children)
 
 
-class _ElementRenderer(intermediate_rendering.DocutilsElementTransformer[_NodeUnion]):
+class _ElementRenderer(intermediate_doc.DocutilsElementTransformer[_NodeUnion]):
     """Render descriptions as C# docstring XML."""
 
     def transform_text(
