+++ aas_core_codegen/golang/description.py
@@ -26,7 +26,6 @@ from aas_core_codegen.golang import (
 )
 from aas_core_codegen.intermediate import (
     doc as intermediate_doc,
-    rendering as intermediate_rendering,
 )
 
 
@@ -56,7 +55,7 @@ class Context:
         self.cls_or_enum = cls_or_enum
 
 
-class _ElementRenderer(intermediate_rendering.DocutilsElementTransformer[str]):
+class _ElementRenderer(intermediate_doc.DocutilsElementTransformer[str]):
     """Render descriptions as a content of a docstring."""
 
     def __init__(self, context: Context) -> None:
