+++ aas_core_codegen/rdf_shacl/_description.py
@@ -7,7 +7,6 @@ from icontract import ensure
 from aas_core_codegen.common import assert_never, assert_union_of_descendants_exhaustive
 from aas_core_codegen.intermediate import (
     doc as intermediate_doc,
-    rendering as intermediate_rendering,
 )
 from aas_core_codegen.rdf_shacl import naming as rdf_shacl_naming
 
@@ -103,7 +102,7 @@ def without_redundant_breaks(tokens: Sequence["TokenUnion"]) -> List["TokenUnion
     return result
 
 
-class Renderer(intermediate_rendering.DocutilsElementTransformer[List["TokenUnion"]]):
+class Renderer(intermediate_doc.DocutilsElementTransformer[List["TokenUnion"]]):
     """Render descriptions as C# docstring XML."""
 
     # fmt: off
